+++ apps/vehicles/src/components/layouts/AppLayout.tsx
@@ -1,21 +1,37 @@
-import { PageHeader, TopBar } from "@io-ipatente/ui";
+import {
+  PageHeader,
+  PageHeaderBreadcrumbsProps,
+  TopBar,
+} from "@io-ipatente/ui";
 import Box from "@mui/material/Box";
-import { ParseKeys } from "i18next";
+import { useRouter } from "next/router";
 import { useTranslation } from "next-i18next";
 import { ReactNode } from "react";
 
-interface AppLayoutProps {
+interface AppLayoutProps extends Partial<PageHeaderBreadcrumbsProps> {
   children: ReactNode;
-  description?: ParseKeys;
-  title: ParseKeys;
+  description?: string;
+  title: string;
 }
 
-const AppLayout = ({ children, description, title }: AppLayoutProps) => {
+const AppLayout = ({
+  breadcrumbs,
+  children,
+  description,
+  onBreadcrumbClick,
+  title,
+}: AppLayoutProps) => {
   const { t } = useTranslation();
+  const router = useRouter();
 
   return (
     <Box display="flex" flexDirection="column" minHeight="100vh">
       <TopBar
+        assistance={{
+          onClick: () => {
+            router.push("/assistance");
+          },
+        }}
         product={{
           logo: "ipatente",
           name: t("topBar.product.name"),
@@ -23,8 +39,10 @@ const AppLayout = ({ children, description, title }: AppLayoutProps) => {
         }}
       />
       <PageHeader
-        description={description ? t(description) : ""}
-        title={t(title)}
+        breadcrumbs={breadcrumbs}
+        description={description}
+        onBreadcrumbClick={onBreadcrumbClick}
+        title={title}
       />
       <Box sx={{ p: 2 }}>{children}</Box>
     </Box>