+++ apps/vehicles/src/components/vehicle-details/VehicleSectionInspections.tsx
@@ -9,12 +9,10 @@ import { useTranslation } from "next-i18next";
 
 export interface VehicleSectionInspectionsProps {
   inspections?: Revisione[];
-  plate: string;
 }
 
 export const VehicleSectionInspections = ({
   inspections = [],
-  plate,
 }: VehicleSectionInspectionsProps) => {
   const { t } = useTranslation();
 
@@ -24,10 +22,14 @@ export const VehicleSectionInspections = ({
         icon={<Icon fontSize="medium" name="documentText" />}
         title={t("vehicleDetails.inspections.title")}
         topContent={
-          <Chip
-            icon={<Icon fontSize="small" name="forbidden" />}
-            label={t("vehicleDetails.inspections.notFound")}
-          />
+          <Typography
+            color="text.secondary"
+            paddingTop={2}
+            textAlign="center"
+            variant="body2"
+          >
+            {t("vehicleDetails.inspections.empty")}
+          </Typography>
         }
       />
     );
@@ -78,20 +80,8 @@ export const VehicleSectionInspections = ({
                   />
                 }
                 label={t(inspectionResultByCode[esitoRevisione.codice].label)}
+                size="small"
               />
-            </ListItem>
-            <ListItem>
-              <Typography
-                display="inline"
-                fontWeight={400}
-                mr={1}
-                variant="body2"
-              >
-                {t("vehicleDetails.inspections.plate")}
-              </Typography>
-              <Typography display="inline" fontWeight={600} variant="body2">
-                {plate}
-              </Typography>
             </ListItem>
             {kmTotali && (
               <ListItem>