+++ packages/ui/lib/components/empty-state/__tests__/EmptyState.test.tsx
@@ -1,0 +1,16 @@
+import { render } from "@testing-library/react";
+import { describe, expect, it } from "vitest";
+
+import { EmptyState } from "../EmptyState";
+
+describe("Test EmptyState Component", () => {
+  it("Should match the snapshot", () => {
+    const comp = render(
+      <EmptyState
+        icon="car1Bold"
+        title="Non sono presenti veicoli registrati"
+      />,
+    );
+    expect(comp).toMatchSnapshot();
+  });
+});