+++ packages/ui/lib/components/index.ts
@@ -1,5 +1,7 @@
+export * from "./assistance-info";
 export * from "./card-carousel";
 export * from "./card-info";
+export * from "./empty-state";
 export * from "./icon";
 export * from "./list-item-action";
 export * from "./logo";