+++ packages/ui/lib/theme.ts
@@ -36,17 +36,6 @@ export const theme = createTheme(
           },
         },
       },
-      MuiButton: {
-        variants: [
-          {
-            style: {
-              "&:hover": {
-                color: colorPrimaryContainedHover,
-              },
-            },
-          },
-        ],
-      },
       MuiCard: {
         styleOverrides: {
           root: {
@@ -71,8 +60,16 @@ export const theme = createTheme(
       MuiListItem: {
         styleOverrides: {
           root: {
-            paddingBottom: muiItaliaTheme.spacing(2),
-            paddingTop: muiItaliaTheme.spacing(2),
+            paddingBottom: muiItaliaTheme.spacing(1.5),
+            paddingTop: muiItaliaTheme.spacing(1.5),
+          },
+        },
+      },
+      MuiListItemButton: {
+        styleOverrides: {
+          root: {
+            paddingBottom: muiItaliaTheme.spacing(1.5),
+            paddingTop: muiItaliaTheme.spacing(1.5),
           },
         },
       },
@@ -96,7 +93,7 @@ export const theme = createTheme(
         selected: "rgba(0, 115, 230, 0.08)",
       },
       background: {
-        default: "#F2F2F2",
+        default: "#FFFFFF",
         pageHeader: colorBackgroundPageHeader,
       },
       divider: "#CBD5E1",