+++ apps/vehicles/CHANGELOG.md
@@ -1,5 +1,47 @@
 # vehicles
 
+## 0.4.0
+
+### Minor Changes
+
+- 2b63b12: Add EmptyState component
+
+### Patch Changes
+
+- Updated dependencies [2b63b12]
+  - @io-ipatente/ui@0.3.0
+
+## 0.3.1
+
+### Patch Changes
+
+- 5dcb84a: Removal of license plate for each inspection
+- Updated dependencies [5dcb84a]
+  - @io-ipatente/ui@0.2.1
+
+## 0.3.0
+
+### Minor Changes
+
+- d912a16: Added infoVeicoli API handler, fix next-auth JWT type augmentation, fix withJWTAuthHandler get session, minor updates
+- 01e16b5: Refinement of ListItemAction
+- 4726cf6: Add vehicle details
+- 782f8fa: Added withJWTAuthHandler bff wrapper
+
+### Patch Changes
+
+- 595f593: Add user's vehicle list
+- 5da5e19: Added generateClientAssertion function for PDND interop client assertion
+- 091d3c4: Updated io-ipatente-vehicles openapi and msw mock data
+- Updated dependencies [595f593]
+- Updated dependencies [5da5e19]
+- Updated dependencies [01e16b5]
+- Updated dependencies [4726cf6]
+- Updated dependencies [9b6a1c7]
+- Updated dependencies [895c829]
+  - @io-ipatente/ui@0.2.0
+  - @io-ipatente/core@0.1.0
+
 ## 0.2.1
 
 ### Patch Changes